Commented the code
Spent way too much time on this. Apart from clues, I did figure out
all of the steps and designed the code myself.

Not sure if this was worth it or not. I did learn quite a bit and it
was a great mind puzzle, but holy hell did it take a long time.